pyproject.toml
[tool.poetry]
name = "interface-grafica"
version = "0.1.0"
description = ""
authors = ["Anderson Rodrigues <eusouanderson@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
mysql-connector-python = "^9.1.0"
pymysql = "^1.1.1"
flask = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
